Revert "Improved xgboost wandb_callback (#1)"
This reverts commit 5559308265a6d0ea4bb61fa6690e63c1ba5ddd3d.
