--- functional_tests/xgboost/xgboost.yea
@@ -1,23 +1,0 @@
-id: 0.xgboost.xgboost
-plugin:
-  - wandb
-command:
-    program: xgboost.py
-depend:
-    requirements:
-        - xgboost>=1.3.0
-        
-assert:
-  - :wandb:runs_len: 1
-  - :wandb:runs[0][config][learner][gradient_booster][name]: gbtree
-  - :wandb:runs[0][config][learner][objective][name]: reg:squarederror
-  - :wandb:runs[0][summary][Feature Importance_table][_type]: table-file
-  - :wandb:runs[0][summary][Feature Importance_table][ncols]: 2
-  - :wandb:runs[0][summary][Feature Importance_table][nrows]: 7
-  - :wandb:runs[0][summary][best_score]: 0.573646
-  - :wandb:runs[0][summary][best_iteration]: 99
-  - :wandb:runs[0][summary][epoch]: 99
-  - :wandb:runs[0][summary][validation_0-rmse.min]: 0.546673
-  - :wandb:runs[0][summary][validation_1-rmse.min]: 0.573646
-  - :wandb:runs[0][exitcode]: 0
-
