+++ wandb/xgboost/__init__.py
@@ -5,6 +5,6 @@ In the future use:
     from wandb.integration.xgboost import wandb_callback
 """
 
-from wandb.integration.xgboost import wandb_callback, WandbCallback
+from wandb.integration.xgboost import wandb_callback
 
-__all__ = ["wandb_callback", "WandbCallback"]
+__all__ = ["wandb_callback"]
